+++ source/dotnet/wholesale-api/Orchestrations/Functions/SettlementReports/SettlementReportOrchestrator.cs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.Wholesale.Orchestrations.Functions.SettlementReports.Activities;
 using Energinet.DataHub.Wholesale.Orchestrations.Functions.SettlementReports.Model;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.DurableTask;
